Use stable versioned references for EC policies
Using main branch for the policy reference means an update to the
ec-policies repo could potentially introduce a breaking change for
RHTAP, perhaps due to the version of the ec-cli, or perhaps for
other reasons, like policy rules being introduced or changes to
existing policy rules.

The `default-v0.4` config file points to a particular branch in the
ec-policies that will remain stable for use with EC v0.4 and hence
for a specific release of RHTAP.

See also https://github.com/enterprise-contract/config/pull/114

Note: As is, this change would be scrubbed the next time
`hack/import-build-definitions` was run. This will be addressed in
upcoming commits.

Ref: https://issues.redhat.com/browse/EC-605